@@ -4,38 +4,33 @@
   "homepage": "web/build",
   "private": true,
   "dependencies": {
-    "@babel/core": "^7.18.13",
-    "@emotion/react": "^11.10.4",
-    "@emotion/styled": "^11.10.4",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.16",
-    "@mui/material": "^5.10.3",
-    "@reduxjs/toolkit": "^1.6.2",
-    "@types/lodash": "^4.14.177",
-    "@types/node": "^16.11.18",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-transition-group": "^4.4.4",
-    "@vitejs/plugin-react": "^1.3.2",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@fortawesome/fontawesome-svg-core": "^6.2.1",
+    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@mui/material": "^5.10.17",
+    "@types/lodash": "^4.14.191",
+    "@types/node": "^18.11.12",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
+    "@types/react-transition-group": "^4.4.5",
+    "@vitejs/plugin-react": "^3.0.0",
     "lodash": "^4.17.21",
-    "mui-nested-menu": "^2.1.2",
-    "react": "^17.0.2",
-    "react-dnd": "^14.0.4",
-    "react-dnd-touch-backend": "^14.1.1",
-    "react-dom": "^17.0.2",
-    "react-markdown": "^7.1.1",
-    "react-redux": "^7.2.6",
+    "mui-nested-menu": "^2.2.0",
+    "react": "^18.2.0",
+    "react-dnd": "^16.0.1",
+    "react-dnd-touch-backend": "^16.0.1",
+    "react-dom": "^18.2.0",
+    "react-markdown": "^8.0.4",
+    "react-redux": "^8.0.5",
     "react-transition-group": "^4.4.5",
-    "redux": "^4.1.2",
-    "redux-devtools": "^3.7.0",
-    "sass": "^1.45.2",
-    "typescript": "^4.5.4",
-    "vite": "^2.9.13"
+    "redux": "^4.2.0",
+    "sass": "^1.56.2"
   },
   "scripts": {
     "start": "vite",
-    "start:game": "vite build --watch",
+    "watch": "vite build --watch",
     "build": "tsc && vite build",
     "preview": "vite preview",
     "format": "prettier --write \"./src/**/*.{ts,tsx,css}\""
@@ -58,15 +53,16 @@
     ]
   },
   "devDependencies": {
+    "@babel/core": "^7.20.5",
+    "@redux-devtools/core": "^3.13.1",
+    "@redux-devtools/instrument": "^2.1.0",
+    "@reduxjs/toolkit": "^1.9.1",
     "@types/react-redux": "^7.1.24",
     "cross-env": "^7.0.3",
-    "csstype": "^2.6.2",
-    "prettier": "^2.5.1",
+    "csstype": "^2.6.21",
+    "prettier": "^2.8.1",
+    "typescript": "^4.9.4",
+    "vite": "^4.0.0",
     "vite-plugin-checker": "^0.4.9"
-  },
-  "pnpm": {
-    "patchedDependencies": {
-      "mui-nested-menu@2.1.2": "patches/mui-nested-menu@2.1.2.patch"
-    }
   }
 }
